alternatives to avoid misuse of word 'ecosystem'
In the English language, the word ecosystem has been introduced around
1935. During those 89 years, it has mostly being used to describe the
system that organisms (that is, biological living systems) form
with their environments. In the last few decades a trend started
in extending the association of this into concepts which are systems,
but far away from being biological; these are rather human-made
organisational structures, business arrangements, or in many cases,
simply mere products, where 'ecosystem' is used with more or less
veiled marketing (greenwashing) intentions. We can have a discussion
about this at deRSE24 or in any other setting, and I'd give you even
more reasons against this abuse.

p.s.: I'm very aware and sorry to jump in the discussion so late..
p.s.: The word has been recognized to be abused already 15 years
 ago within the environmental sciences:
https://www.cambridge.org/core/journals/environmental-conservation/article/on-the-use-and-misuse-of-the-term-ecosystem/A57FCEBC08E64CFC7409FD40F6DA97FE

Signed-off-by: Michele Martone <michelemartone@users.sourceforge.net>